Moving IPCAWrapper into its own class for cleanliness
